+++ include/encoding/BooleanDecoder.hpp
@@ -12,11 +12,11 @@ protected:
 	uint8_t* bufferend;
 	uint8_t offset = 0;
 public:
-	BooleanDecoder(uint8_t* buffer, uint32_t size);
+	BooleanDecoder(uint8_t* buffer, uint64_t size);
 	~BooleanDecoder() {}
 	uint8_t* nextValue();
 	bool get();
-	uint8_t getValueSize() { return 0; }
+	uint32_t getValueSize() { return 0; }
 };
