+++ include/encoding/RleDecoder.hpp
@@ -18,13 +18,13 @@ private:
 	uint8_t repeat_value = 0;
 	uint8_t bitwidth = 0;
 public:
-	RleDecoder(uint8_t* buffer, uint32_t& maxsize, uint8_t bitwidth);
+	RleDecoder(uint8_t* buffer, uint64_t& maxsize, uint8_t bitwidth);
 	RleDecoder(uint8_t* buffer, uint8_t* bufferend, uint8_t bitwidth);
 	RleDecoder(); // not to be used
 	~RleDecoder() {}
 	bool get(uint8_t& val);
 	uint8_t* nextValue() { return nullptr; }
-	uint8_t getValueSize() { return 0; }
+	uint32_t getValueSize() { return 0; }
 };
