+++ include/util/ThriftUtil.hpp
@@ -17,9 +17,9 @@ namespace util {
 
 
 template <typename T>
-T* thrift_deserialize(uint8_t* mem, uint32_t& size) {
+T* thrift_deserialize(uint8_t* mem, uint64_t& size) {
     T* obj = new T();
-    boost::shared_ptr<TMemoryBuffer> tmem_transport(new TMemoryBuffer(const_cast<uint8_t*>(mem), size));
+    boost::shared_ptr<TMemoryBuffer> tmem_transport(new TMemoryBuffer(const_cast<uint8_t*>(mem), uint32_t(size)));
     TCompactProtocolFactoryT<TMemoryBuffer> tproto_factory;
     boost::shared_ptr<TProtocol> tproto = tproto_factory.getProtocol(tmem_transport);
     size = obj->read(tproto.get());
