+++ src/ParquetColumn.cpp
@@ -9,11 +9,11 @@ ParquetColumn::ParquetColumn(uint8_t* mem, parquet::thriftschema::ColumnMetaData
 		: mem(mem), metadata(metadata), schema(schema), dict_page(nullptr) {
 	if (dict_mem != nullptr) {
 		// format guarantees that only one dictionary page exists
-		uint32_t dictheader_size = metadata.total_uncompressed_size;
+		uint64_t dictheader_size = metadata.total_uncompressed_size;
 		auto* dictpageheader = util::thrift_deserialize<parquet::thriftschema::PageHeader>(dict_mem, dictheader_size);
 		dict_page = new ParquetDictionaryPage(dict_mem+dictheader_size, dictpageheader->uncompressed_page_size, dictpageheader->dictionary_page_header.num_values, schema);
 	}
-	uint32_t header_size = metadata.total_uncompressed_size;
+	uint64_t header_size = metadata.total_uncompressed_size;
 	mem_end = mem + metadata.total_uncompressed_size;
 	auto* pageheader = util::thrift_deserialize<parquet::thriftschema::PageHeader>(mem, header_size);
 	assert(pageheader->type == parquet::thriftschema::PageType::DATA_PAGE);
@@ -27,7 +27,7 @@ void ParquetColumn::nextPage() {
 		cur_page = nullptr;
 		return;
 	}
-	uint32_t header_size = mem_end-mem;
+	uint64_t header_size = mem_end-mem;
 	auto* pageheader = util::thrift_deserialize<parquet::thriftschema::PageHeader>(mem, header_size);
 	assert(pageheader->type == parquet::thriftschema::PageType::DATA_PAGE);
 	cur_page = new ParquetDataPage(mem+header_size, pageheader->uncompressed_page_size, pageheader->data_page_header, schema, dict_page);
