+++ src/ParquetFile.cpp
@@ -18,7 +18,7 @@ ParquetFile::ParquetFile(const std::string& filename) : filename(filename) {
 	// check magic number
 	if (memcmp(file_mem, "PAR1", 4) != 0) throw "not a parquet file";
 	if (memcmp(file_mem+file_size-4, "PAR1", 4) != 0) throw "not a parquet file";
-	uint32_t footersize = *reinterpret_cast<uint32_t*>(file_mem+file_size-8);
+	uint64_t footersize = *reinterpret_cast<uint32_t*>(file_mem+file_size-8);
     uint8_t* buf = file_mem + (file_size-8-footersize);
     filemetadata = util::thrift_deserialize<parquet::thriftschema::FileMetaData>(buf, footersize);
 	schema = readSchema();
