+++ test/encoding/RleDecoderTest.cpp
@@ -13,7 +13,7 @@ TEST(RleDecoderTest, BitPack) {
     buffer[5] = 0b10001000;
     buffer[6] = 0b11000110;
     buffer[7] = 0b11111010;
-    uint32_t maxsize = 8;
+    uint64_t maxsize = 8;
     RleDecoder d(buffer, maxsize, (uint8_t)3);
     uint8_t val;
     for (uint8_t i=0; i < 8; i++) {
@@ -28,7 +28,7 @@ TEST(RleDecoderTest, RunLengthEncoding) {
     reinterpret_cast<uint32_t*>(buffer)[0] = 2; // length of following data
     buffer[4] = 0b00000110; // repeated 3 times
     buffer[5] = (uint8_t)5;
-    uint32_t maxsize = 8;
+    uint64_t maxsize = 8;
     RleDecoder d(buffer, maxsize, (uint8_t)3);
     uint8_t val;
     for (uint8_t i=0; i < 3; i++) {
